Rather than trying to restrict the drift adjustment into a sane range, just stop applying it if it's huge or tiny.
git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@656 ed5b42cb-e716-0410-a449-f6a68f950b19

+++ src/java/com/threerings/media/FrameManager.java
@@ -221,12 +221,12 @@ public abstract class FrameManager
      * Returns the highest drift ratio our timer has seen. 1.0 means no drift (and is what we return
      * if we're not using a CalibratingTimer)
      */
-    public double getMaxTimerDriftRatio ()
+    public float getMaxTimerDriftRatio ()
     {
     	if (_timer instanceof CalibratingTimer) {
     	    return ((CalibratingTimer)_timer).getMaxDriftRatio();
     	} else {
-    		return 1.0;
+    		return 1.0F;
     	}
     }
